Move proofofphonenumberverfication code to latte
